backend/requirements.txt
black==25.9.0
boto3==1.40.69
botocore==1.40.69
celery==5.5.3
dj-database-url==3.0.1
Django==5.2.7
django-environ==0.11.2
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
flake8-bugbear==25.10.21
flake8-comprehensions==3.17.0
gunicorn==23.0.0
Jinja2==3.1.3
platformdirs==4.5.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg-binary==3.2.12
PyJWT==2.10.1
pytokens==0.3.0
pytz==2024.2
PyYAML==6.0.3
rapidfuzz==3.9.0
redis==7.0.1
requests==2.31.0
trio==0.25.0
trio-websocket==0.11.1
tzdata==2025.2
urllib3==2.2.1
virtualenv==20.35.4
pytest>=8.0
pytest-django>=4.9